Updating filesystem manipulation files to remove warnings
